packages/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.1",
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyproj>=3.7.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis[asyncio]>=7.2.1",
    "shapely>=2.1.2",
    "simplekml>=1.3.6",
    "uvicorn[standard]>=0.41.0",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
